requirements.txt
azure-core==1.30.1
azure-search-documents==11.4.0
openai==1.23.2
python-dotenv==1.0.1
pyodbc==5.1.0